[Guardian] Separate combined ceremony state
