+++ crates/hashi-types/src/guardian/log/message.rs
@@ -28,7 +28,6 @@ use crate::committee::CommitteeSignature;
 use crate::guardian::GuardianError;
 use crate::guardian::GuardianInfo;
 use crate::guardian::GuardianPubKey;
-use crate::guardian::GuardianResult;
 use crate::guardian::KPEncryptedShares;
 use crate::guardian::LimiterState;
 use crate::guardian::NitroAttestation;
@@ -254,39 +253,6 @@ impl CeremonyLogMessage {
 
     fn object_key_pattern(&self, session_id: &str) -> ObjectKeyPattern {
         ObjectKeyPattern::Fixed(self.object_key(session_id))
-    }
-}
-
-/// The current ceremony result together with its latest encrypted KP share state.
-#[derive(Clone, Debug, PartialEq)]
-pub struct CeremonyState {
-    pub secret_sharing_instance: SecretSharingInstance,
-    pub btc_master_pubkey: BitcoinPubkey,
-    pub cert_seq: u64,
-    pub encrypted_shares: KPEncryptedShares,
-}
-
-impl CeremonyState {
-    /// Combine a ceremony log message with the KP share-state message for its
-    /// resulting secret-sharing instance.
-    pub fn new(
-        ceremony: CeremonyLogMessage,
-        kp_share_state: KpShareStateLogMessage,
-    ) -> GuardianResult<Self> {
-        let (secret_sharing_instance, btc_master_pubkey) = ceremony.into_instance_and_pubkey();
-        if kp_share_state.sharing_seq != secret_sharing_instance.sharing_seq() {
-            return Err(GuardianError::InvalidInputs(format!(
-                "kp-shares sharing_seq ({}) differs from ceremony sharing_seq ({})",
-                kp_share_state.sharing_seq,
-                secret_sharing_instance.sharing_seq()
-            )));
-        }
-        Ok(Self {
-            secret_sharing_instance,
-            btc_master_pubkey,
-            cert_seq: kp_share_state.cert_seq,
-            encrypted_shares: kp_share_state.encrypted_shares,
-        })
     }
 }
